Remove .mo files from repo.
This was tried before but apparently gave test problems, which is weird.
Let's try again though.  It indeed fails.
But let's install zest.pocompile in .travis.yml and run it.
That should fix it.
And we add 'zope_i18n_compile_mo_files true' in the instance,
so they are compile locally too.