Fix multipart error on search
This error connection happens randomly and breaks a thread connection
permanently.

```
o.a.s.s.SolrRequestParsers Couldn't get multipart parts in order to
delete them => java.io.IOException: Missing initial multi part boundary
  at org.eclipse.jetty.util.MultiPartInputStreamParser.parse(MultiPartInputStreamParser.java:606)
java.io.IOException: Missing initial multi part boundary
```

The fix resets the connection and completely recreates it from within
the manager. (A simple connection reset is not sufficient.)

There is no evidence of what causes the issue but a reindex with many
content items breaks the connection of the thread that was doing the
reindex in a reliable way. This solution fixes this scenario.

- Also update black version