WIP Fix reindex if an attribute is removed from an object
If an attribute becomes None, then the previous code would not have
updated the attribute value in the index and the previous value would
have lingered indefinitely.

This is a second variety fix which instead of unindexing the whole
object, just removes all attributes that are not in the data.
Caveat: currently effective only on string and text field types.