WIP Prevent wildcard expressions for stopwords in simple expressions (#368)
- add stopwords to registry
- add stopwords_case_insensitive option
- support function for getting the stopwords
- cache to optimize the parsing of stopwords.txt

This transforms the (term AND term*) expression for stopwords, removing
the wildcard expression. Such an expression would never match any
documents, because solr won't remove the wildcard term, but the
stopword will be missing from the index. This workaround does that with
no side effects, as stopwords would be ignored by solr anyway.

Both case sensitive and case insensitive stopword processing is
supported, this depends on the solr schema, and must be set accordingly.